Update dashboard robot model
- Move model into own directory.
- Add colors to model.
- Add dashboard robot view model.
